Integrated grape functionality into Schrodinger * `grape`, `NormPSU` and `CoherentSubspacesPSU` are exported * can now pass `Optim.Options` to `grape` * moved inner product functions to new file * slightly changed interface again
