+++ src/control/grape.jl
@@ -1,7 +1,7 @@
-using Schrodinger, Optim
+using Optim: Options
+
 abstract type ObjectiveFunction end
-include("normpsu.jl")
-include("coherentsubspaces.jl")
+const IntCol = Union{AbstractVector{Int},IntSet,Set{Int},NTuple{N,Int} where N}
 
 function calc_fprops!(U,X,D,V,u,δt,Hd,Hc,H,u_last)
     # If the control amplitudes u did not change, return
@@ -14,7 +14,7 @@ function calc_fprops!(U,X,D,V,u,δt,Hd,Hc,H,u_last)
             H .+= u[(j-1)*m+k].*Hc[k]
         end
         # Calculate U and store eigenvectors and eigenvalues of -δt*H
-        Schrodinger.expim!(U[j],Hermitian(scale!(H,-δt)),D[j],V[j],X[1])
+        expim!(U[j],Hermitian(scale!(H,-δt)),D[j],V[j],X[1])
     end
     # Calculate forward propagators (cumulative product of U)
     copy!(X[1],U[1])
@@ -55,74 +55,16 @@ function _Jmathermprod!(J,cisDj,Dj,δt)
     return nothing
 end
 
-function grape(O::ObjectiveFunction,u_init)
+function grape(O::ObjectiveFunction,u_init::Array,opt::Options=Options())
     # Build OnceDifferentiable object from ObjectiveFunction object
-    od = OnceDifferentiable(O,(fp,u)->O(Val{:gradient}(),fp,u),zeros(O.u_last))
-    # Optimization options
-    #opt = Optim.Options(g_tol = 1E-9)
+    f(u) = objective(O,u)
+    g!(fp,u) = gradient!(O,fp,u)
+    fg!(fp,u) = (g!(fp,u); f(u))
+    seed = ones(O.u_last)
+    od = OnceDifferentiable(f,g!,fg!,1.0,similar(seed),seed,copy(seed),[1],[1])
     # Run optimization
-    res = optimize(od,vec(u_init.'),ConjugateGradient())
+    res = optimize(od,vec(u_init.'),ConjugateGradient(),opt)
     # Reshape final control amplitude matrix
     uf = reshape(Optim.minimizer(res),size(u_init,2),size(u_init,1)).'
-    return uf, O.X[end], res
-end
-
-function opt_pihalfx(n)
-    Hd = qzero(2)
-    Hc = [π*σx, π*σy]
-    t = 1
-    #u_init = rand(n,1) .- 0.5
-    u_init = zeros(n,2)
-    Ut = expm(-1im*π*σx/4)
-    # Create objective function type
-    O = NormPSU(Ut,Hd,Hc,t,n)
-    grape(O,u_init)
-end
-
-function opt_hadamard(n)
-    Hd = σz
-    Hc = [σx]
-    t = 10.0
-    #u_init = rand(n,1) .- 0.5
-    u_init = zeros(n,1)
-    Ut = Operator([1/√2 1/√2; 1/√2 -1/√2])
-    # Create objective function type
-    O = NormPSU(Ut,Hd,Hc,t,n)
-    grape(O,u_init)
-end
-
-function opt_3lvlNOT(n)
-    Δ = 2π*(-400) # anharmonicity
-    Hd = Δ*Operator(basis(3,2)) # drift Hamiltonian
-    Hc = [create(3)/2+destroy(3)/2, im*create(3)/2-im*destroy(3)/2]
-    t = 3
-    u_init = [-Δ*Schrodinger.gaussianpulse.(linspace(-t/2,t/2,n),[[t/2,t,0,0,pi]]) linspace(-Δ/2,Δ/2,n)]
-    Ut = Operator([0 1 0; 1 0 0; 0 0 1]) # 3lvl NOT gate
-    # Create objective function type
-    O = CoherentSubspaces(Ut,1:2,Hd,Hc,t,n) # care only about computational subspace
-    grape(O,u_init)
-end
-
-function opt_2Q_QFT(n)
-    Hd = 0.5 * (σx⊗σx + σy⊗σy + σz⊗σz)
-    Hc = [0.5*σx⊗σ0, 0.5*σy⊗σ0, 0.5*σ0⊗σx, 0.5*σ0⊗σy]
-    t = 6.0
-    u_init = zeros(n,4)
-    #u_init = rand(n,4) .- 0.5
-    Ut = Operator([0.5  0.5    0.5  0.5
-                   0.5  0.5im -0.5 -0.5im
-                   0.5 -0.5    0.5 -0.5
-                   0.5 -0.5im -0.5  0.5im],
-                  (2,2))
-    # Create objective function type
-    O = NormPSU(Ut,Hd,Hc,t,n)
-    grape(O,u_init)
-end
-
-function plotgrape(tf,uf)
-    figure()
-    step(linspace(0,tf,size(uf,1)+1),[uf[1:1,:]; uf])
-    legend(["Control $i" for i in 1:size(uf,2)])
-    tight_layout()
-    grid()
+    return uf, Operator(O.X[end],O.dims), res
 end
